requirements.txt
python-geoip-python3
requests
clint